fix failing test cases

+++ hub/htypes.py
@@ -28,8 +28,9 @@ HTYPE_CONFIGURATIONS: Dict[str, Dict] = {
         "sample_compression": "png",
         "chunk_compression": UNCOMPRESSED,
     },
+    # TODO: hub auto needs dtype as int64
     "class_label": {
-        "dtype": "int64",
+        "dtype": "uint32",
         "class_names": [],
     },
     "bbox": {"dtype": "float32"},
